requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1
PyYAML==6.0.2
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
sqlalchemy==2.0.36
python-dateutil==2.9.0.post0
pdfplumber==0.11.4
rapidfuzz==3.10.1
loguru==0.7.3
